# server/src/docs/swagger.yaml
definitions:
...
  models.File:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      id:
        type: integer
      object_key:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Persona:
    properties:
      created_at:
        type: string
      id:
        type: integer
      persona_description:
        type: string
      persona_title:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
    required:
    - persona_description
    - persona_title
    type: object
...
  models.SubTask:
    properties:
      actions:
        type: string
      actions_state:
        type: string
      created_at:
        type: string
      id:
        type: integer
      sub_task_description:
        type: string
      sub_task_name:
        type: string
      task_id:
        type: integer
      updated_at:
        type: string
    required:
    - sub_task_description
    - sub_task_name
    - task_id
    type: object
  models.Tag:
    properties:
      created_at:
        type: string
      files:
        items:
          $ref: '#/definitions/models.File'
        type: array
      id:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      id:
        type: integer
      personas:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      task_description:
        type: string
      task_name:
        type: string
      updated_at:
        type: string
    required:
    - task_description
    - task_name
    type: object
...
info:
  contact:
    email: oduneye.d@northeastern.edu and dupuguntla@northeastern.edu
    name: David Oduneye and Akshay Dupuguntla
  description: Backend Server for Legacy App
  title: Legacy API
  version: "1.0"
paths:
...
  /api/subtasks/:
    get:
      description: Returns all subtasks
      operationId: get-all-subtasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SubTask'
            type: array
        "404":
          description: Failed to fetch subtasks
          schema:
            type: string
      summary: Gets all subtasks
      tags:
      - subtask
    post:
      consumes:
      - application/json
      description: Creates a subtask
      operationId: create-subtask
      parameters:
      - description: SubTask
        in: body
        name: subtask
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubTask'
        "400":
          description: Failed to create subtask
          schema:
            type: string
      summary: Creates a subtask
      tags:
      - subtask
  /api/subtasks/{sid}:
    delete:
      description: Deletes a subtask
      operationId: delete-subtask
      parameters:
      - description: SubTaskID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subtask successfully deleted
          schema:
            type: string
        "404":
          description: Failed to delete subtask
          schema:
            type: string
      summary: Deletes a subtask
      tags:
      - subtask
    get:
      description: Returns a subtask from subtask id
      operationId: get-subtask
      parameters:
      - description: SubTaskID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubTask'
        "404":
          description: Failed to fetch subtask
          schema:
            type: string
      summary: Gets a subtask from subtask id
      tags:
      - subtask
    patch:
      consumes:
      - application/json
      description: Updates a subtask
      operationId: update-subtask
      parameters:
      - description: SubTaskID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubTask'
        "400":
          description: Failed to update subtask
          schema:
            type: string
      summary: Updates a subtask
      tags:
      - subtask
  /api/subtasks/{sid}/actions:
    get:
      description: Returns all actions from subtask id
      operationId: get-actions
      parameters:
      - description: SubTaskID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SubTask'
            type: array
        "404":
          description: Failed to fetch subtask
          schema:
            type: string
      summary: Gets all actions from subtask id
      tags:
      - subtask
  /api/tasks/:
    get:
      description: Returns all tasks
      operationId: get-all-tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: Failed to fetch tasks
          schema:
            type: string
      summary: Gets all tasks
      tags:
      - task
    post:
      consumes:
      - application/json
      description: Creates a task
      operationId: create-task
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Failed to create task
          schema:
            type: string
      summary: Creates a task
      tags:
      - task
  /api/tasks/{tid}:
    delete:
      description: Deletes a task
      operationId: delete-task
      parameters:
      - description: TaskID
        in: path
        name: tid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully deleted
          schema:
            type: string
        "404":
          description: Failed to delete task
          schema:
            type: string
      summary: Deletes a task
      tags:
      - task
    get:
      description: Returns a task from task id
      operationId: get-task
      parameters:
      - description: TaskID
        in: path
        name: tid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Failed to fetch task
          schema:
            type: string
      summary: Gets a task from task id
      tags:
      - task
    patch:
      consumes:
      - application/json
      description: Updates a task
      operationId: update-task
      parameters:
      - description: TaskID
        in: path
        name: tid
        required: true
        type: string
      - description: Task
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Failed to update task
          schema:
            type: string
      summary: Updates a task
      tags:
      - task
  /api/tasks/{tid}/subtasks:
    get:
      description: Returns all subtasks of a task
      operationId: get-all-subtasks-of-task
      parameters:
      - description: TaskID
        in: path
        name: tid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: Failed to fetch subtasks
          schema:
            type: string
      summary: Gets all subtasks of a task
      tags:
      - task
  /api/tasks/{uid}/user:
    get:
      description: Returns all tasks of a user
      operationId: get-all-user-tasks
      parameters:
      - description: UserID
        in: path
        name: uid
        required: true
        type: string
      - description: Tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: Failed to fetch tasks
          schema:
            type: string
      summary: Gets all tasks of a user
      tags:
      - task
...
swagger: "2.0"